[RSDK-9981] Add a skeleton of a module (#1)
This is enough that I can use it as a local module and get a module and a service both up and running, and get a bunch of `NotImplementedError`s if I actually try using it.

Next steps will be to add in some actual functionality, but this PR is already big enough on its own. Don't bother reviewing commit-by-commit: it's easier to see the whole thing together. A lot of this was auto-generated through the Viam CLI; I've written very little code so far.


* run 'viam module generate'

* run 'ruff format main.py'

* line-wrap comments to 80 characters

* line wrap long argument lists, too

* start a setup section in the Makefile

* add pyinstaller to requirements.txt, alphabetize and pin versions

* add a Makefile target to create an executable

* hook into the auto-generated setup.sh

* use autogenerated build.sh instead of making our own in the Makefile

* tweak: newline at end of file

* tweak: use pyinstaller from the local path

* debugging