Support -rc releases with Github Actions (#2)
I'm unsure how to try out any of this without first merging it. and I don't want to try it out until there's more functionality in the module itself. but my hope is that because this is copied and pasted from the checklist, it's probably right-ish even without being able to try it out yet.


* copy deploy action from Triton module

* get rid of boilerplate for private repos

* remove unnecessary comment: you can use a -rc1 version if you don't want to deploy to production

* rename deploy.yml -> publish.yml

* never mind: copy and paste from the checklist